--- Cha/CalculatorModel.java
@@ -1,29 +1,0 @@
-package campus_console_project_leader_side2.Cha;
-
-public class CalculatorModel {
-    private double result;
-
-    public void add(double num1, double num2) {
-        result = num1 + num2;
-    }
-
-    public void subtract(double num1, double num2) {
-        result = num1 - num2;
-    }
-
-    public void multiply(double num1, double num2) {
-        result = num1 * num2;
-    }
-
-    public void divide(double num1, double num2) {
-        if (num2 != 0) {
-            result = num1 / num2;
-        } else {
-            result = 0;
-        }
-    }
-
-    public double getResult() {
-        return result;
-    }
-}
